requirements.txt
transformers
protobuf
sentencepiece
chromadb
sentence-transformers
python-dotenv
torch
huggingface_hub
bitsandbytes
accelerate
